// src/types/anilist.ts
export enum MediaFormat {
  Tv = "TV",
  Tv_short = "TV_SHORT",
  Movie = "MOVIE",
  Special = "SPECIAL",
  Ova = "OVA",
  Ona = "ONA",
  Music = "MUSIC",
  Manga = "MANGA",
  Novel = "NOVEL",
  One_shot = "ONE_SHOT",
}
/** The format the language was released in */
export enum LanguagesDetectFormat {
  jpn = "ja",
  eng = "en",
  vie = "vi",
  chi_sim = "zh",
}
/** The format the language was released in */
export enum RevertLanguagesDetectFormat {
  ja = "jpn",
  en = "eng",
  vi = "vie",
  zh = "chi_sim",
}
/** The current releasing status of the media */
export enum MediaStatus {
  Finished = "FINISHED",
  Releasing = "RELEASING",
  Not_yet_released = "NOT_YET_RELEASED",
  Cancelled = "CANCELLED",
  Hiatus = "HIATUS",
}
/** Date object that allows for incomplete date values (fuzzy) */
export interface FuzzyDate {
  /** Numeric Year (2017)*/
  year: Maybe<number>;
  /** Numeric Month (3)*/
  month: Maybe<number>;
  /** Numeric Day (24)*/
  day: Maybe<number>;
}

export enum MediaSeason {
  Winter = "WINTER",
  Spring = "SPRING",
  Summer = "SUMMER",
  Fall = "FALL",
}
// ...
export enum CharacterRole {
  Main = "MAIN",
  Supporting = "SUPPORTING",
  Background = "BACKGROUND",
}
// ...
export enum MediaSort {
  Id = "ID",
  Id_desc = "ID_DESC",
  Title_romaji = "TITLE_ROMAJI",
  Title_romaji_desc = "TITLE_ROMAJI_DESC",
  Title_english = "TITLE_ENGLISH",
  Title_english_desc = "TITLE_ENGLISH_DESC",
  Title_native = "TITLE_NATIVE",
  Title_native_desc = "TITLE_NATIVE_DESC",
  Type = "TYPE",
  Type_desc = "TYPE_DESC",
  Format = "FORMAT",
  Format_desc = "FORMAT_DESC",
  Start_date = "START_DATE",
  Start_date_desc = "START_DATE_DESC",
  End_date = "END_DATE",
  End_date_desc = "END_DATE_DESC",
  Score = "SCORE",
  Score_desc = "SCORE_DESC",
  Popularity = "POPULARITY",
  Popularity_desc = "POPULARITY_DESC",
  Trending = "TRENDING",
  Trending_desc = "TRENDING_DESC",
  Episodes = "EPISODES",
  Episodes_desc = "EPISODES_DESC",
  Duration = "DURATION",
  Duration_desc = "DURATION_DESC",
  Status = "STATUS",
  Status_desc = "STATUS_DESC",
  Chapters = "CHAPTERS",
  Chapters_desc = "CHAPTERS_DESC",
  Volumes = "VOLUMES",
  Volumes_desc = "VOLUMES_DESC",
  Updated_at = "UPDATED_AT",
  Updated_at_desc = "UPDATED_AT_DESC",
  Search_match = "SEARCH_MATCH",
  Favourites = "FAVOURITES",
  Favourites_desc = "FAVOURITES_DESC",
}